Improve handling of setting roundtime
Move it to a function, since we are setting it from multiple places. At
the same time remove the need to restart the round, as we can just set
the roundtime for the current round that is ongoing.

Also the setting of the gamerules causes exceptions to be thrown
sometimes, since the timelimit can change before gamerules are properly
loaded into sourcemod. Could not find any fix for this, so we just
disable the setting of roundtime until we are later in the loading of
things. A bit hacky, but seems to work.